requirements.txt
django
djangorestframework
django-url-filter
furl
psycopg2
python-memcached
gunicorn
pytz
raven
requests